Coalesce duplicate implementations of KernelData.
Replace three identical inner classes with a single top-level class
adjacent to the the abstract super-class.
